Swap events for commands; Commands are now classes
The title tells most of the story. Commands are now classes rather than
just functions, giving a bit more context to the whole situation and
allowing us to extend functionality a bit easier. They are also (mostly)
self documenting now as they take the given docstring for the class and
use that as the help text.